services/onboarding-orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx[http2]>=0.27.0
asyncpg>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
